refactor: view 패키지 구분
