+++ src/main/java/step2/view/ResultView.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.view;
 
 public class ResultView {
     private static void printMessage(String message) {
